requirements.txt
asgiref==3.7.2
Django==5.0.1
django-mailer==2.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==22.6.0
gunicorn==21.2.0
inflection==0.5.1
lockfile==0.12.2
packaging==23.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
result==0.16.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
whitenoise==6.6.0
